requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.6
Django==4.0
django-crispy-forms==1.13.0
sqlparse==0.4.2
